Revert "Revert "revert-commit-fix-report""
This reverts commit fb564f1aafa719949c979f769deac5cda09890f3.

+++ src/modules/report/components/QuantityRevenueItem.jsx
@@ -12,13 +12,10 @@ import Swal from "sweetalert2";
 export default function QuantityRevenueItem({
   detail,
   categories,
-  remainingCategories,
   updateTotalAmount = () => {},
   onChange = () => {},
   onRemove = () => {},
-  onCategorySelect,
 }) {
-  // console.log(remainingCategories);
   // Hàm để tính tổng tiền
   const [quantity, setQuantity] = useState(detail?.quantity || "");
   const [price, setPrice] = useState(detail?.price || "");
@@ -44,56 +41,40 @@ export default function QuantityRevenueItem({
 
   const [unit, setUnit] = useState(detail?.unit || "");
   const handleSelectCategory = async (event) => {
-    debugger;
     const category = event.target.value;
     const selectedCategory = categories.find((el) => el.name === category);
-    if (remainingCategories.length <= 1) {
-      onCategorySelect(false);
+    if (selectedCategory) {
       setUnit(selectedCategory.unit);
       setPrice(selectedCategory.price);
-      onChange({
-        ...detail,
-        unit: selectedCategory.unit,
-        price: selectedCategory.price,
-        category,
-      });
-      return;
     }
-    const validateCategory = onCategorySelect(selectedCategory);
-    if (validateCategory) {
-      setUnit(selectedCategory.unit);
-      setPrice(selectedCategory.price);
-      onChange({
-        ...detail,
-        unit: selectedCategory.unit,
-        price: selectedCategory.price,
-        category,
-      });
-    }
+    onChange({
+      ...detail,
+      unit: selectedCategory.unit,
+      price: selectedCategory.price,
+      category,
+    });
   };
-  //xóa item
+
   const deleteDiv = async (id) => {
     try {
-      // const result = await Swal.fire({
-      //   title: "Bạn chắc chắn muốn xóa? ",
-      //   text: "Hạng mục này sẽ bị xóa vĩnh viễn!",
-      //   icon: "warning",
-      //   showCancelButton: true,
-      //   confirmButtonColor: "#3085d6",
-      //   cancelButtonColor: "#d33",
-      //   confirmButtonText: "Xóa hạng mục",
-      //   cancelButtonText: "Hủy bỏ",
-      // });
-      // if (result.isConfirmed) {
-      //   Swal.fire({
-      //     title: "Đã xóa!",
-      //     text: "Hạng mục đã được xóa thành công.",
-      //     icon: "success",
-      //   });
-      //   onRemove(detail);
-      // }
-      debugger;
-      onRemove(detail);
+      const result = await Swal.fire({
+        title: "Bạn chắc chắn muốn xóa? ",
+        text: "Hạng mục này sẽ bị xóa vĩnh viễn!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Xóa hạng mục",
+        cancelButtonText: "Hủy bỏ",
+      });
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Đã xóa!",
+          text: "Hạng mục đã được xóa thành công.",
+          icon: "success",
+        });
+        onRemove(detail);
+      }
     } catch (error) {}
   };
 
@@ -121,6 +102,7 @@ export default function QuantityRevenueItem({
               labelId="demo-simple-select-label"
               id="demo-simple-select"
               value={detail?.category}
+              // label="Hạng mục*"
               label={
                 <span>
                   Hạng mục<span style={{ color: "red" }}>*</span>
@@ -130,12 +112,12 @@ export default function QuantityRevenueItem({
               size="small"
               sx={{ display: "flex", width: "450px" }}
             >
-              {remainingCategories?.map((category) => (
+              {categories?.map((category) => (
                 <MenuItem
                   key={category.id + "_" + category.value}
-                  value={detail?.category ? detail?.category : category.name}
+                  value={category.name}
                 >
-                  {detail?.category ? detail?.category : category.name}
+                  {category.name}
                 </MenuItem>
               ))}
             </Select>
